requirements.txt
# -i https://pypi.org/simple
# numpy==2.3.2; python_version >= '3.11'
# scipy==1.16.1; python_version >= '3.11'
# pyarrow==21.0.0; python_version >= '3.9'

python-dotenv==1.1.1; python_version >= '3.9'
# tzdata
click==8.2.1

numpy==1.26.4
scipy==1.16.1; python_version >= '3.11'

beautifulsoup4==4.13.4
selenium==4.25.0
regex==2025.9.1
nltk==3.9.1

requests==2.32.5
urllib3==2.5.0
# pymupdf==1.24.10
PyMuPDF==1.24.10

# pgvector==0.7.4
psycopg2-binary==2.9.10; python_version >= '3.8'
pandas==2.3.2; python_version >= '3.9'

# scikit-learn==1.7.1; python_version >= '3.10'
# fastembed==0.7.3; python_full_version >= '3.9.0'
faiss-cpu==1.8.0.post1
# huggingface-hub==0.34.4; python_full_version >= '3.8.0'
openai==1.106.1; python_version >= '3.8'

streamlit==1.49.1; python_version >= '3.9' and python_full_version != '3.9.7'
flask==3.1.2; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'